@@ -7,8 +7,8 @@ buildscript {
         
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.6.1'
-        classpath 'org.greenrobot:greendao-gradle-plugin:3.2.2'
+        classpath 'com.android.tools.build:gradle:4.0.1'
+
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
@@ -30,11 +30,7 @@ task clean(type: Delete) {
 ext {
 
     // SDK And Tools
-    minSdkVersion = 19
-    targetSdkVersion = 28
-    compileSdkVersion = 28
-
-    //Dependencies
-    supportLibraryVersion = '28.0.0'
-    constraintVersion = '1.1.3'
+    minSdkVersion = 21
+    targetSdkVersion = 29
+    compileSdkVersion = 29
 }
