@@ -6,8 +6,8 @@ package com.hjc.reader.constant;
  * @Description: EventBus消息标识常量类
  */
 public class EventCode {
-    public static final int A = 0x111111;
-    public static final int B = 0x222222;
-    public static final int C = 0x333333;
-    public static final int D = 0x444444;
+    public static final int LOGIN_CODE = 0x111111;
+    public static final int EDIT_LINK_CODE = 0x222222;
+    public static final int CLICK_TAG = 0x333333;
+    public static final int SEARCH_RESULT = 0x444444;
 }
