@@ -1,36 +1,33 @@
 package com.hjc.reader.ui.film.child;
 
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import android.app.Activity;
+import android.os.Bundle;
 import android.view.View;
+import android.widget.ImageView;
 
-import com.blankj.utilcode.util.ToastUtils;
+import androidx.core.app.ActivityOptionsCompat;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.recyclerview.widget.LinearLayoutManager;
+
 import com.chad.library.adapter.base.BaseQuickAdapter;
+import com.hjc.baselib.fragment.BaseMvmLazyFragment;
 import com.hjc.reader.R;
-import com.hjc.reader.base.fragment.BaseLazyFragment;
-import com.hjc.reader.http.RetrofitHelper;
-import com.hjc.reader.http.helper.RxHelper;
-import com.hjc.reader.http.observer.BaseProgressObserver;
-import com.hjc.reader.model.response.MovieHotBean;
-import com.hjc.reader.model.response.MovieItemBean;
+import com.hjc.reader.bean.response.MovieItemBean;
+import com.hjc.reader.constant.RoutePath;
+import com.hjc.reader.databinding.FragmentHotMovieBinding;
 import com.hjc.reader.ui.film.adapter.HotMovieAdapter;
-import com.scwang.smartrefresh.layout.SmartRefreshLayout;
+import com.hjc.reader.utils.helper.RouteManager;
+import com.hjc.reader.viewmodel.film.HotMovieViewModel;
 
 import java.util.List;
 
-import butterknife.BindView;
 
 /**
  * @Author: HJC
  * @Date: 2019/1/21 11:29
  * @Description: 热映电影页面
  */
-public class HotMovieFragment extends BaseLazyFragment {
-    @BindView(R.id.smart_refresh_layout)
-    SmartRefreshLayout smartRefreshLayout;
-    @BindView(R.id.rv_list)
-    RecyclerView rvList;
-
+public class HotMovieFragment extends BaseMvmLazyFragment<FragmentHotMovieBinding, HotMovieViewModel> {
     private HotMovieAdapter mAdapter;
 
     public static HotMovieFragment newInstance() {
@@ -43,71 +40,68 @@ public class HotMovieFragment extends BaseLazyFragment {
     }
 
     @Override
+    protected HotMovieViewModel getViewModel() {
+        return new ViewModelProvider(this).get(HotMovieViewModel.class);
+    }
+
+    @Override
+    protected int getBindingVariable() {
+        return 0;
+    }
+
+    @Override
     public void initView() {
+        setLoadSir(mBindingView.smartRefreshLayout);
+
         LinearLayoutManager manager = new LinearLayoutManager(mContext);
-        rvList.setLayoutManager(manager);
+        mBindingView.rvList.setLayoutManager(manager);
 
         mAdapter = new HotMovieAdapter(null);
-        rvList.setAdapter(mAdapter);
+        mBindingView.rvList.setAdapter(mAdapter);
 
-        mAdapter.openLoadAnimation(BaseQuickAdapter.SLIDEIN_BOTTOM);
+        mAdapter.setAnimationWithDefault(BaseQuickAdapter.AnimationType.SlideInBottom);
     }
 
     @Override
     public void initData() {
-        getHotMovieData(true);
+        mViewModel.loadHotMovieData(true);
     }
 
-    /**
-     * 获取热映榜电影数据
-     *
-     * @param isShow 是否显示loading
-     */
-    private void getHotMovieData(boolean isShow) {
-        RetrofitHelper.getInstance().getMTimeService()
-                .getHotFilm()
-                .compose(RxHelper.bind(this))
-                .subscribe(new BaseProgressObserver<MovieHotBean>(getChildFragmentManager(), isShow) {
-                    @Override
-                    public void onSuccess(MovieHotBean result) {
-                        smartRefreshLayout.finishRefresh();
-                        if (result != null) {
-                            parseMovieData(result);
-                        } else {
-                            ToastUtils.showShort("未获取到数据");
-                        }
-                    }
-
-                    @Override
-                    public void onFailure(String errorMsg) {
-                        super.onFailure(errorMsg);
-                        smartRefreshLayout.finishRefresh();
-                    }
-                });
-    }
-
-
-    /**
-     * 解析豆瓣电影数据
-     *
-     * @param result 电影数据对应的bean
-     */
-    private void parseMovieData(MovieHotBean result) {
-        List<MovieItemBean> movieList = result.getMs();
-        if (movieList != null && movieList.size() > 0) {
-            mAdapter.setNewData(movieList);
-        } else {
-            ToastUtils.showShort("数据解析异常");
-        }
+    @Override
+    protected void observeLiveData() {
+        mViewModel.getHotMovieLiveData().observe(this, data -> {
+            mBindingView.smartRefreshLayout.finishRefresh();
+            mAdapter.setNewInstance(data);
+        });
     }
 
     @Override
     public void addListeners() {
-        smartRefreshLayout.setOnRefreshListener(refreshLayout -> getHotMovieData(false));
+        mBindingView.smartRefreshLayout.setOnRefreshListener(refreshLayout -> {
+            mViewModel.loadHotMovieData(false);
+        });
+
+        mAdapter.setOnItemClickListener((adapter, view, position) -> {
+            List<MovieItemBean> dataList = mAdapter.getData();
+            MovieItemBean bean = dataList.get(position);
+
+            ImageView ivCover = view.findViewById(R.id.iv_cover);
+
+            Bundle bundle = new Bundle();
+            bundle.putSerializable("bean", bean);
+            ActivityOptionsCompat compat = ActivityOptionsCompat.makeSceneTransitionAnimation((Activity) mContext, ivCover, mContext.getResources().getString(R.string.transition_movie_img));
+            RouteManager.jumpWithScene(mContext, RoutePath.URL_MOVIE_DETAIL, bundle, compat);
+        });
+    }
+
+    @Override
+    protected void onRetryBtnClick(View v) {
+        mViewModel.loadHotMovieData(true);
     }
 
     @Override
     public void onSingleClick(View v) {
 
     }
+
 }
