@@ -1,15 +1,16 @@
 package com.hjc.reader.ui.gank;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.design.widget.TabLayout;
-import android.support.v4.app.Fragment;
-import android.support.v4.view.ViewPager;
 import android.view.View;
 
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+
+import com.hjc.baselib.fragment.BaseMvmFragment;
+import com.hjc.baselib.viewmodel.CommonViewModel;
 import com.hjc.reader.R;
 import com.hjc.reader.adapter.MyViewPagerAdapter;
-import com.hjc.reader.base.fragment.BaseFragment;
+import com.hjc.reader.databinding.FragmentTab2Binding;
 import com.hjc.reader.ui.gank.child.GankFragment;
 import com.hjc.reader.ui.gank.child.RecommendFragment;
 import com.hjc.reader.ui.gank.child.WelfareFragment;
@@ -17,25 +18,16 @@ import com.hjc.reader.ui.gank.child.WelfareFragment;
 import java.util.ArrayList;
 import java.util.List;
 
-import butterknife.BindView;
-
 /**
  * @Author: HJC
  * @Date: 2019/1/21 16:10
  * @Description: 干货模块
  */
-public class Tab2Fragment extends BaseFragment {
-
-    @BindView(R.id.tab_layout)
-    TabLayout tabLayout;
-    @BindView(R.id.view_pager)
-    ViewPager viewPager;
-
-    private String titles[] = new String[]{"每日推荐", "福利社区", "干货定制"};
+public class Tab2Fragment extends BaseMvmFragment<FragmentTab2Binding, CommonViewModel> {
+    private String[] titles = new String[]{"每日推荐", "福利社区", "干货定制"};
 
     public static Tab2Fragment newInstance() {
-        Tab2Fragment fragment = new Tab2Fragment();
-        return fragment;
+        return new Tab2Fragment();
     }
 
     @Override
@@ -44,8 +36,13 @@ public class Tab2Fragment extends BaseFragment {
     }
 
     @Override
-    public void initView() {
+    protected CommonViewModel getViewModel() {
+        return null;
+    }
 
+    @Override
+    protected int getBindingVariable() {
+        return 0;
     }
 
     @Override
@@ -61,9 +58,9 @@ public class Tab2Fragment extends BaseFragment {
         fragments.add(gankFragment);
 
         MyViewPagerAdapter adapter = new MyViewPagerAdapter(getChildFragmentManager(), fragments, titles);
-        viewPager.setAdapter(adapter);
-        viewPager.setOffscreenPageLimit(3);
-        tabLayout.setupWithViewPager(viewPager);
+        mBindingView.viewPager.setAdapter(adapter);
+        mBindingView.viewPager.setOffscreenPageLimit(3);
+        mBindingView.tabLayout.setupWithViewPager(mBindingView.viewPager);
     }
 
     @Override
