@@ -1,61 +1,42 @@
 package com.hjc.reader.ui.gank.child;
 
-import android.annotation.SuppressLint;
-import android.graphics.Rect;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import android.graphics.Color;
 import android.view.View;
-import android.widget.ImageView;
 
-import com.blankj.utilcode.util.ToastUtils;
+import androidx.annotation.NonNull;
+import androidx.lifecycle.ViewModelProvider;
+import androidx.recyclerview.widget.GridLayoutManager;
+
 import com.chad.library.adapter.base.BaseQuickAdapter;
+import com.hitomi.tilibrary.style.index.NumberIndexIndicator;
+import com.hitomi.tilibrary.style.progress.ProgressPieIndicator;
+import com.hitomi.tilibrary.transfer.TransferConfig;
+import com.hitomi.tilibrary.transfer.Transferee;
+import com.hjc.baselib.fragment.BaseMvmLazyFragment;
 import com.hjc.reader.R;
-import com.hjc.reader.base.event.EventManager;
-import com.hjc.reader.base.event.MessageEvent;
-import com.hjc.reader.base.fragment.BaseLazyFragment;
-import com.hjc.reader.constant.EventCode;
-import com.hjc.reader.http.RetrofitHelper;
-import com.hjc.reader.http.helper.RxHelper;
-import com.hjc.reader.http.helper.RxSchedulers;
-import com.hjc.reader.http.observer.BaseProgressObserver;
-import com.hjc.reader.model.ImageViewInfo;
-import com.hjc.reader.model.response.GankIOBean;
+import com.hjc.reader.bean.response.GankDayBean;
+import com.hjc.reader.databinding.FragmentWelfareBinding;
 import com.hjc.reader.ui.gank.adapter.WelfareAdapter;
-import com.hjc.reader.utils.SchemeUtils;
-import com.hjc.reader.utils.ViewUtils;
-import com.scwang.smartrefresh.layout.SmartRefreshLayout;
+import com.hjc.reader.viewmodel.gank.WelfareViewModel;
 import com.scwang.smartrefresh.layout.api.RefreshLayout;
 import com.scwang.smartrefresh.layout.listener.OnRefreshLoadMoreListener;
-
-import org.greenrobot.eventbus.Subscribe;
-import org.greenrobot.eventbus.ThreadMode;
+import com.vansz.glideimageloader.GlideImageLoader;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.TimeUnit;
-
-import butterknife.BindView;
-import io.reactivex.Observable;
-import io.reactivex.functions.Consumer;
 
 /**
  * @Author: HJC
  * @Date: 2019/1/21 11:29
  * @Description: 福利社区页面
  */
-public class WelfareFragment extends BaseLazyFragment {
-
-    @BindView(R.id.smart_refresh_layout)
-    SmartRefreshLayout smartRefreshLayout;
-    @BindView(R.id.rv_welfare)
-    RecyclerView rvWelfare;
+public class WelfareFragment extends BaseMvmLazyFragment<FragmentWelfareBinding, WelfareViewModel> {
 
     private WelfareAdapter mAdapter;
-    private int page = 1;
-    private GridLayoutManager mGridLayoutManager;
 
-    private ArrayList<ImageViewInfo> imageViewInfoList;
+    private Transferee transferee;
+
+    private int mPage = 1;
 
     public static WelfareFragment newInstance() {
         return new WelfareFragment();
@@ -67,103 +48,101 @@ public class WelfareFragment extends BaseLazyFragment {
     }
 
     @Override
+    protected WelfareViewModel getViewModel() {
+        return new ViewModelProvider(this).get(WelfareViewModel.class);
+    }
+
+    @Override
+    protected int getBindingVariable() {
+        return 0;
+    }
+
+    @Override
     public void initView() {
-        mGridLayoutManager = new GridLayoutManager(mContext, 2);
-        rvWelfare.setLayoutManager(mGridLayoutManager);
+        setLoadSir(mBindingView.smartRefreshLayout);
 
-        mAdapter = new WelfareAdapter(null);
-        rvWelfare.setAdapter(mAdapter);
+        GridLayoutManager manager = new GridLayoutManager(mContext, 2);
+        mBindingView.rvWelfare.setLayoutManager(manager);
 
-        mAdapter.openLoadAnimation(BaseQuickAdapter.SCALEIN);
+        mAdapter = new WelfareAdapter();
+        mBindingView.rvWelfare.setAdapter(mAdapter);
+
+        mAdapter.setAnimationWithDefault(BaseQuickAdapter.AnimationType.ScaleIn);
     }
 
     @Override
     public void initData() {
-        EventManager.register(this);
-        getWelfareData(true);
+        transferee = Transferee.getDefault(mContext);
+
+        mViewModel.loadWelfareList(mPage, true);
     }
 
-    /**
-     * 获取福利数据
-     */
-    private void getWelfareData(boolean isShow) {
-        RetrofitHelper.getInstance().getGankIOService()
-                .getGankIoData("福利", 20, page)
-                .compose(RxHelper.bind(this))
-                .subscribe(new BaseProgressObserver<GankIOBean>(getChildFragmentManager(), isShow) {
-                    @Override
-                    public void onSuccess(GankIOBean result) {
-                        smartRefreshLayout.finishRefresh();
-                        smartRefreshLayout.finishLoadMore();
-                        if (result != null) {
-                            parseWelfareData(result);
-                        } else {
-                            ToastUtils.showShort("未获取到数据");
-                        }
-                    }
+    @Override
+    protected void observeLiveData() {
+        mViewModel.getWelfareLiveData().observe(this, data -> {
+            mBindingView.smartRefreshLayout.finishRefresh();
+            mBindingView.smartRefreshLayout.finishLoadMore();
 
-                    @Override
-                    public void onFailure(String errorMsg) {
-                        super.onFailure(errorMsg);
-                        smartRefreshLayout.finishRefresh();
-                        smartRefreshLayout.finishLoadMore();
-                    }
-                });
-    }
-
-    /**
-     * 解析福利数据
-     *
-     * @param gankIOBean 福利对应的bean
-     */
-    private void parseWelfareData(GankIOBean gankIOBean) {
-        List<GankIOBean.ResultsBean> resultList = gankIOBean.getResults();
-        if (resultList != null && resultList.size() > 0) {
-            if (page == 1) {
-                mAdapter.setNewData(resultList);
-            } else {
-                mAdapter.addData(resultList);
+            if (data != null) {
+                if (mPage == 1) {
+                    mAdapter.setNewInstance(data);
+                } else {
+                    mAdapter.addData(data);
+                }
             }
-        }
+        });
     }
 
     @Override
     public void addListeners() {
-        smartRefreshLayout.setOnRefreshLoadMoreListener(new OnRefreshLoadMoreListener() {
+        mBindingView.smartRefreshLayout.setOnRefreshLoadMoreListener(new OnRefreshLoadMoreListener() {
             @Override
             public void onLoadMore(@NonNull RefreshLayout refreshLayout) {
-                page++;
-                getWelfareData(false);
+                mPage++;
+                mViewModel.loadWelfareList(mPage, false);
             }
 
             @Override
             public void onRefresh(@NonNull RefreshLayout refreshLayout) {
-                page = 1;
-                getWelfareData(false);
+                mPage = 1;
+                mViewModel.loadWelfareList(mPage, false);
             }
         });
 
         mAdapter.setOnItemClickListener((adapter, view, position) -> {
-            imageViewInfoList = new ArrayList<>();
-            List<GankIOBean.ResultsBean> dataList = adapter.getData();
-            for (GankIOBean.ResultsBean bean : dataList) {
-                ImageViewInfo viewInfo = new ImageViewInfo(bean.getUrl());
-                imageViewInfoList.add(viewInfo);
+            List<GankDayBean> dataList = mAdapter.getData();
+            List<String> imagesList = new ArrayList<>();
+            for (GankDayBean bean : dataList) {
+                imagesList.add(bean.getUrl());
             }
 
-            //获取view位置坐标信息
-            int firstVisiblePosition = mGridLayoutManager.findFirstVisibleItemPosition();
-            int lastVisiblePosition = mGridLayoutManager.findLastVisibleItemPosition();
-            for (int i = firstVisiblePosition; i < lastVisiblePosition + 1; i++) {
-                View itemView = mGridLayoutManager.findViewByPosition(i);
-                if (itemView != null) {
-                    ImageView imageView = itemView.findViewById(R.id.iv_pic);
-                    Rect rect = ViewUtils.computeBound(imageView);
-                    imageViewInfoList.get(i).setRect(rect);
-                }
-            }
-            SchemeUtils.jumpToGalley(mContext, imageViewInfoList, position, firstVisiblePosition, lastVisiblePosition);
+            TransferConfig config = TransferConfig.build()
+                    .setSourceUrlList(imagesList) // 资源 url 集合, String 格式
+                    .setMissPlaceHolder(R.mipmap.img_default_meizi) // 资源加载前的占位图
+                    .setErrorPlaceHolder(R.mipmap.img_default_meizi) // 资源加载错误后的占位图
+                    .setProgressIndicator(new ProgressPieIndicator()) // 资源加载进度指示器, 可以实现 IProgressIndicator 扩展
+                    .setIndexIndicator(new NumberIndexIndicator()) // 资源数量索引指示器，可以实现 IIndexIndicator 扩展
+                    .setImageLoader(GlideImageLoader.with(mContext)) // 图片加载器，可以实现 ImageLoader 扩展
+                    .setBackgroundColor(Color.parseColor("#000000")) // 背景色
+                    .setDuration(300) // 开启、关闭、手势拖拽关闭、显示、扩散消失等动画时长
+                    .setOffscreenPageLimit(2) // 第一次初始化或者切换页面时预加载资源的数量，与 justLoadHitImage 属性冲突，默认为 1
+                    .setNowThumbnailIndex(position) // 缩略图在图组中的索引
+                    .enableJustLoadHitPage(true) // 是否只加载当前显示在屏幕中的的资源，默认关闭
+                    .enableDragClose(true) // 是否开启下拉手势关闭，默认开启
+                    .enableDragHide(false) // 下拉拖拽关闭时，是否先隐藏页面上除主视图以外的其他视图，默认开启
+                    .enableDragPause(false) // 下拉拖拽关闭时，如果当前是视频，是否暂停播放，默认关闭
+                    .enableHideThumb(false) // 是否开启当 transferee 打开时，隐藏缩略图, 默认关闭
+                    .enableScrollingWithPageChange(false) // 是否启动列表随着页面的切换而滚动你的列表，默认关闭
+                    .bindRecyclerView(mBindingView.rvWelfare, R.id.iv_pic);  // 绑定一个 RecyclerView， 所有绑定方法只能调用一个
+
+            transferee.apply(config).show();
         });
+    }
+
+    @Override
+    protected void onRetryBtnClick(View v) {
+        mPage = 1;
+        mViewModel.loadWelfareList(mPage, true);
     }
 
     @Override
@@ -174,31 +153,6 @@ public class WelfareFragment extends BaseLazyFragment {
     @Override
     public void onDestroy() {
         super.onDestroy();
-        EventManager.unregister(this);
-    }
-
-    @SuppressLint("CheckResult")
-    @Subscribe(threadMode = ThreadMode.MAIN)
-    public void handlerEvent(MessageEvent<Integer> messageEvent) {
-        if (messageEvent.getCode() == EventCode.A) {
-            Integer currentPosition = messageEvent.getData();
-            mGridLayoutManager.scrollToPositionWithOffset(currentPosition, 0);
-
-            //延时500ms,不然ItemView获取不到
-            Observable.timer(500, TimeUnit.MILLISECONDS)
-                    .compose(RxSchedulers.ioToMain())
-                    .subscribe(new Consumer<Long>() {
-                        @Override
-                        public void accept(Long aLong) throws Exception {
-                            View itemView = mGridLayoutManager.findViewByPosition(currentPosition);
-                            if (itemView != null) {
-                                ImageView imageView = itemView.findViewById(R.id.iv_pic);
-                                Rect rect = ViewUtils.computeBound(imageView);
-                                imageViewInfoList.get(currentPosition).setRect(rect);
-                                EventManager.sendEvent(new MessageEvent<>(EventCode.B, imageViewInfoList));
-                            }
-                        }
-                    });
-        }
+        transferee.destroy();
     }
 }
