@@ -1,138 +1,84 @@
 package com.hjc.reader.ui.login;
 
+
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.design.widget.TextInputEditText;
 import android.view.View;
-import android.widget.Button;
 
+import androidx.annotation.Nullable;
+import androidx.lifecycle.ViewModelProvider;
+
+import com.alibaba.android.arouter.facade.annotation.Route;
 import com.blankj.utilcode.util.KeyboardUtils;
-import com.blankj.utilcode.util.StringUtils;
-import com.blankj.utilcode.util.ToastUtils;
+import com.hjc.baselib.activity.BaseMvmActivity;
 import com.hjc.reader.R;
-import com.hjc.reader.base.activity.BaseActivity;
-import com.hjc.reader.http.RetrofitHelper;
-import com.hjc.reader.http.helper.RxHelper;
-import com.hjc.reader.http.observer.BaseProgressObserver;
-import com.hjc.reader.model.response.LoginBean;
-import com.hjc.reader.widget.TitleBar;
+import com.hjc.reader.constant.RoutePath;
+import com.hjc.reader.databinding.ActivityRegisterBinding;
+import com.hjc.reader.utils.helper.RouteManager;
+import com.hjc.reader.viewmodel.login.RegisterViewModel;
 
-import butterknife.BindView;
 
 /**
  * @Author: HJC
- * @Date: 2019/2/13 15:25
+ * @Date: 2020/5/14 15:27
  * @Description: 注册页面
  */
-public class RegisterActivity extends BaseActivity {
-    @BindView(R.id.title_bar)
-    TitleBar titleBar;
-    @BindView(R.id.et_username)
-    TextInputEditText etUsername;
-    @BindView(R.id.et_password)
-    TextInputEditText etPassword;
-    @BindView(R.id.et_confirm_password)
-    TextInputEditText etConfirmPassword;
-    @BindView(R.id.btn_register)
-    Button btnRegister;
+@Route(path = RoutePath.URL_REGISTER)
+public class RegisterActivity extends BaseMvmActivity<ActivityRegisterBinding, RegisterViewModel> {
 
     @Override
-    public int getLayoutId() {
+    protected int getLayoutId() {
         return R.layout.activity_register;
     }
 
     @Override
-    public void initView() {
-
+    protected RegisterViewModel getViewModel() {
+        return new ViewModelProvider(this).get(RegisterViewModel.class);
     }
 
     @Override
-    public void initData(Bundle savedInstanceState) {
-
+    protected int getBindingVariable() {
+        return 0;
     }
 
     @Override
-    public void addListeners() {
-        btnRegister.setOnClickListener(this);
+    protected void initData(@Nullable Bundle savedInstanceState) {
+        mBindingView.setRegisterViewModel(mViewModel);
+    }
 
-        titleBar.setOnViewClickListener(new TitleBar.onViewClick() {
-            @Override
-            public void leftClick(View view) {
+    @Override
+    protected void observeLiveData() {
+        mViewModel.getRegisterData().observe(this, isRegister -> {
+            if (isRegister){
                 KeyboardUtils.hideSoftInput(RegisterActivity.this);
+
+                Intent intent = new Intent();
+                intent.putExtra("username", mBindingView.etUsername.getText());
+                intent.putExtra("password", mBindingView.etPassword.getText());
+                setResult(1000, intent);
                 finish();
-            }
-
-            @Override
-            public void rightClick(View view) {
-
             }
         });
     }
 
     @Override
-    public void onSingleClick(View v) {
-        switch (v.getId()) {
-            case R.id.btn_register:
-                register();
-                break;
-        }
+    protected void addListeners() {
+        mBindingView.setOnClickListener(this);
+
+        mBindingView.titleBar.setOnViewLeftClickListener(view -> finish());
     }
 
-    private void register() {
-        String username = etUsername.getText().toString().trim();
-        String password = etPassword.getText().toString().trim();
-        String confirmPassword = etConfirmPassword.getText().toString().trim();
+    @Override
+    protected void onSingleClick(View v) {
+        switch (v.getId()) {
+            case R.id.btn_register:
+                mViewModel.register();
+                break;
 
-        if (StringUtils.isEmpty(username)) {
-            ToastUtils.showShort("请输入用户名");
-            return;
+            case R.id.tv_register:
+                RouteManager.jump(RoutePath.URL_LOGIN);
+                break;
         }
-
-        if (username.length() < 6) {
-            ToastUtils.showShort("用户名长度至少为6位");
-            return;
-        }
-
-        if (StringUtils.isEmpty(password)) {
-            ToastUtils.showShort("请输入密码");
-            return;
-        }
-
-        if (password.length() < 6) {
-            ToastUtils.showShort("密码长度至少为6位");
-            return;
-        }
-
-        if (StringUtils.isEmpty(password)) {
-            ToastUtils.showShort("请输入确认密码");
-            return;
-        }
-
-        if (!password.equals(confirmPassword)) {
-            ToastUtils.showShort("两次密码输入不一致");
-            return;
-        }
-
-        RetrofitHelper.getInstance().getWanAndroidService()
-                .register(username, password, confirmPassword)
-                .compose(RxHelper.bind(this))
-                .subscribe(new BaseProgressObserver<LoginBean>(getSupportFragmentManager()) {
-                    @Override
-                    public void onSuccess(LoginBean result) {
-                        if (result != null) {
-                            ToastUtils.showShort("注册成功");
-                            KeyboardUtils.hideSoftInput(RegisterActivity.this);
-
-                            Intent intent = new Intent();
-                            intent.putExtra("username", username);
-                            intent.putExtra("password", password);
-                            setResult(1000, intent);
-                            finish();
-                        } else {
-                            ToastUtils.showShort("注册失败,请稍后重试");
-                        }
-                    }
-                });
     }
 
 }
